+++ src/main/java/com/lmx1/enhancedpollution/Utils/TileUtils.java
@@ -9,10 +9,10 @@ public class TileUtils {
     //public List<TileEntity> masters = new ArrayList<TileEntity>();
 
     //Returns amount of pollution that Tile produces
-    public static float getPollution(TileEntity tile){
+    public static double getPollution(TileEntity tile){
         //System.out.println(tile.getClass().getName());
         if(tile instanceof TileEntityFurnace){
-            return ((TileEntityFurnace) tile).isBurning() ? 50f : 0.0f;
+            return ((TileEntityFurnace) tile).isBurning() ? 50d : 0.0d;
         }
         /*if(tile.getClass().getName().startsWith("blusunrize.immersiveengineering.common")){
             return IEMultiBlockIntegration.getPollution(tile);
